day03: speed up execution time
My original solution for part 1 used Sets of coordinates and it was
fairly speedy (~0.5 s on my laptop).  To solve part 2 I decided to keep
track of the steps toward each coordinate by changing the code to use
HashMaps instead, which degraded performance by ~3x (1.5 s for part 1).
Thinking the problem through a bit more, I realised that with
a small-ish number of intersections (compared to the total number of
coordinates in the path of the wire) it would probably be cheaper to
calculate the number of steps taken to reach each one instead, taking
advantage of the faster original implementation.  The final version
takes around 1 s for both parts.

Convert HashMaps to Sets of coordinates.  Make Coord a structured type
and make Direction a vector-like type to simplify the calculation of
movement.  Change the solution to part 2 to explicitly calculate the
number of steps taken to reach each intersection through each wire by
stepping through the wire one leg at a time and checking whether the
intersection is part of the leg.