day13: solve puzzles
The _actual_ fastest way to solve the second part is to replace every
tile on the same row as the original paddle with more paddles and then
just keep feeding the computer zeros until the game finishes.  This is
definitely cheating, but also perversely fun.

This is not the fun solution.